+++ test/vitest/geolocation/get-current-position.spec.ts
@@ -34,7 +34,9 @@ describe("Geolocation.getCurrentPosition()", () => {
     vi.useFakeTimers({ shouldAdvanceTime: true });
     vi.setSystemTime(startTime);
 
-    ({ geolocation, locationServices, user } = createAPIs({ acquireDelay: 0 }));
+    ({ geolocation, locationServices, user } = createAPIs({
+      locationServicesParams: { acquireDelay: 0 },
+    }));
 
     successCallback = vi.fn();
     errorCallback = vi.fn();